Remove use of Regex to parse DNS records
The current solution isn't error-proof either, but it's slightly more
reliable.

This change is needed because TXT DNS records can, and do, contain ;
inside them, like DKIM TXT records. This made the library stop early
when parsing, and made the rest of the code fail.

This solution is somewhat more reliable. It will first filter the HTML
page to find the corresponding `<script>` that contains the records
data, and then it will do string search to find where the `records`
declaration starts and ends. This is still error-prone, but less so, and
it will support complex DNS records with ;.